# Support insert replace completion capability