Could be related with https://github.com/redhat-developer/vscode-java/pull/2343

Signed-off-by: sheche <sheche@microsoft.com>