@jdneo 2022-03-31
@Eskibear This change will return a [InsertReplaceEdit](https://microsoft.github.io/language-server-protocol/specifications/lsp/3.17/specification/#insertReplaceEdit), will this change affect intellicode?