@jdneo 2022-04-02
This video illustrates how the capability works.


https://user-images.githubusercontent.com/6193897/161367615-ba1451b8-3e59-464b-88ef-43156d77deda.mp4



The default behavior depends on the client setting.

In VS Code it's `editor.suggest.insertMode`. and the mode can be changed by pressing `Shift`