@rgrunber 2022-04-06
Doing a very rough performance comparison by triggering completion 10 times yields very little difference for me.

|Run #|current time (ms)|PR time (ms)|
|--|---|---|
|1|266|173|
|2|135|72|
|3|84|58|
|4|53|68|
|5|55|48|
|6|46|43|
|7|39|42|
|8|38|50|
|9|47|37|
|10|37|34|