@jdneo 2022-04-07
I got a similar result on my Windows 11. Trigger completion after character `S` for 10 times:

|Run #|current time (ms)|PR time (ms)|
|--|---|---|
|1|2424|2531|
|2|182|299|
|3|151|133|
|4|143|129|
|5|143|152|
|6|137|146|
|7|198|125|
|8|101|178|
|9|91|134|
|10|90|102|