@jdneo 2022-04-07
By the way if we merge this pull request, I think [this one](https://github.com/redhat-developer/vscode-java/pull/2343) can be discard then. And even we might consider totally remove the setting `java.completion.overwrite` at VS Code client side.