@rgrunber 2022-04-07
As you mentioned in https://github.com/redhat-developer/vscode-java/pull/2343#issuecomment-1071928905 . With this change, the default behaviour of "overwrite" (`java.completion.overwrite=true`) would become "insert" (`editor.suggest.insertMode=insert`).

Are we ok with this ? Should we try to preserve the old behaviour with some kind of language-specific setting of `editor.suggest.insertMode` ?