@jdneo 2022-04-08
Since the survey results show that there is no strong preference between them. Personally, I think it's fine to change the default behavior. On the other hand, another challenge might be how we can let users know that they can press `Shift` to change toggle it.