@jdneo 2022-04-12
> I'm guessing we can remove java.completion.overwrite from the vscode-java client

Sure, I'll update the [PR](https://github.com/redhat-developer/vscode-java/pull/2343) to remove it