@rgrunber 2022-04-13
It looks like the testcase here is failing pretty consistently on the builds, although passing locally https://ci.eclipse.org/ls/job/jdt-ls-master/1298/testReport/org.eclipse.jdt.ls.core.internal.handlers/CompletionInsertReplaceCapabilityTest/testCompletion_InsertReplaceEdit/ .