@testforstephen 2022-04-13
It should be related with running JDK version and its src.zip.

For some JDK distribution, the JDK binary classes doesn't include the exact method parameter names, the code completion handler would need to use the JDK source code to find the parameter name. If the JDK src is missing, it will use the default name like "arg0"...