@jdneo 2022-04-13
+1, it should be related with whether the source of JDK is available or not. I'll make another PR to fix the test.